+++ app/head.tsx
@@ -1,0 +1,9 @@
+export default function Head() {
+  return (
+    <>
+      <title>Queres ser o meu padrinho?</title>
+      <meta name="description" content="Queres ser o meu padrinho?" />
+      <link rel="icon" href="/favicon.ico" />
+    </>
+  )
+}
